requirements.txt
amqp==2.6.1
asgiref==3.3.1
astroid==2.4.2
awesome-slugify==1.6.5
billiard==3.6.3.0
braintree==3.59.0
cairocffi==1.2.0
CairoSVG==2.5.1
celery==4.4.7
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.0.11
django-crispy-forms==1.10.0
django-localflavor==3.0.1
django-parler==2.2
django-prometheus==2.1.0
django-rosetta==0.9.5
easy-thumbnails==2.7.1
flower==0.9.7
html5lib==1.1
humanize==3.2.0
idna==2.10
importlib-metadata==3.4.0
isort==5.7.0
kombu==4.6.11
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.1.0
polib==1.1.0
prometheus-client==0.8.0
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
pycparser==2.20
pylint==2.6.0
Pyphen==0.10.0
python-stdnum==1.16
pytz==2020.5
redis==3.5.3
regex==2020.11.13
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
toml==0.10.2
tornado==6.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
Unidecode==0.4.21
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
WeasyPrint==52.2
webencodings==0.5.1
wrapt==1.12.1
zipp==3.4.0
